+++ scarletio/utils/function_tools.py
@@ -101,7 +101,12 @@ def copy_func(old):
     new : `functions`
         The new created function.
     """
-    new = FunctionType(old.__code__, old.__globals__, name=old.__name__, argdefs=old.__defaults__,
-        closure=old.__closure__)
+    new = FunctionType(
+        old.__code__,
+        old.__globals__,
+        name = old.__name__,
+        argdefs = old.__defaults__,
+        closure = old.__closure__,
+    )
     new.__kwdefaults__ = old.__kwdefaults__
     return new